life_coachAi_SalmaZaher/requirements.txt
openai
streamlit
python-dotenv
pytz
google-api-python-client
google-auth
google-auth-oauthlib
langchain
